use existing rename-variable
